requirements.txt
# ── Core Deep Learning ───────────────────────────────────────
tensorflow>=2.13.0          # GPU build auto-selected if CUDA is available
# tensorflow-gpu>=2.13.0    # uncomment this & remove 'tensorflow' for GPU-only

# ── Data & Numerics ──────────────────────────────────────────
numpy>=1.24.0
pillow>=10.0.0              # image loading / resizing in predict.py
requests>=2.31.0            # URL image download in predict.py

# ── Machine Learning Metrics ─────────────────────────────────
scikit-learn>=1.3.0         # confusion_matrix, classification_report

# ── Visualisation ────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Notebook ─────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.25.0

# ── Optional: Interactive Web UI ─────────────────────────────
# gradio>=4.0.0             # uncomment to enable the --ui flag in predict.py
